+++ packages/vitepress-theme-project-trans/src/config.ts
@@ -17,6 +17,34 @@ import { defineConfig } from 'vitepress'
 import { generateSidebar } from './sidebar'
 import { useThemeContext } from './utils/themeContext'
 
+import fs from 'fs';
+import path from 'path';
+
+// 从文件系统读取 Markdown 文件内容
+function readMarkdownFileContent(filePath: string): string {
+  if (fs.existsSync(filePath)) {
+    return fs.readFileSync(filePath, 'utf-8');
+  }
+  return '';
+}
+
+// 统计文档的字数函数
+function countWords(content: string): number {
+  const cleanedContent = content
+    .replace(/```[\s\S]*?```/g, '') // 移除代码块
+    .replace(/!\[.*?\]\(.*?\)/g, '') // 移除图片链接
+    .replace(/\[.*?\]\(.*?\)/g, '') // 移除普通链接
+    .replace(/<\/?[^>]+(>|$)/g, '') // 移除 HTML 标签
+    .replace(/[.,\/#!$%\^&\*;:{}=\-_`~()]/g, '') // 移除标点符号
+    .replace(/\s+/g, ' ') // 将多余的空格归为一个空格
+    .trim(); // 去除首尾空格
+
+  const chineseCharacters = cleanedContent.match(/[\u4e00-\u9fff\uff01-\uffe5]/g) || [];
+  const words = cleanedContent.split(/\s+/).filter(Boolean);
+
+  return chineseCharacters.length + words.length;
+}
+
 // https://vitepress.dev/reference/site-config
 function genConfig() {
   const themeConfig = useThemeContext()
@@ -196,7 +224,25 @@ function genConfig() {
         ],
       },
     },
-  })
+    transformPageData(pageData) {
+      // 构建 Markdown 文件路径
+      const markdownFile = `${pageData.relativePath}`;
+      const filePath = path.join(process.cwd(), 'docs', markdownFile);
+
+      // 从文件系统读取文件内容
+      const content = readMarkdownFileContent(filePath);
+
+      // 统计字数并插入到 Frontmatter
+      const wordCount = countWords(content);
+
+      return {
+        frontmatter: {
+          ...pageData.frontmatter,
+          wordCount, // 将字数写入 Frontmatter
+        },
+      };
+    },
+  });
 }
 
-export default genConfig
+export default genConfig;
